🔧 refactor(bot): substituir herança de CrawJUD por ClassBot e reorganizar código
Alterações realizadas para melhorar a estrutura do código e a legibilidade. A classe Pauta agora herda de ClassBot em vez de CrawJUD, promovendo uma melhor organização. Além disso, um novo arquivo JSON foi adicionado para categorizar cidades do Amazonas, e arquivos antigos foram removidos para limpar a base de código.

Files Changed:
- *crawjud/bots/pje/pauta.py*: substituição da herança de CrawJUD por ClassBot.
- *crawjud/common/bot.py*: adição da classe ClassBot e métodos relacionados.
- *crawjud/common/cities_amazonas.json*: novo arquivo para categorizar cidades do Amazonas.
- *crawjud/core/__init__.py*: remoção do arquivo para limpar a base de código.
- *crawjud/core/data_formatters/__init__.py*: remoção do arquivo para limpar a base de código.
